+++ app/client/src/utils/autoLayout/positionUtils.test.ts
@@ -3,11 +3,10 @@ import {
   Positioning,
   ResponsiveBehavior,
 } from "utils/autoLayout/constants";
-import { FlexLayer } from "./autoLayoutTypes";
+import { AlignmentInfo, FlexLayer, Row } from "./autoLayoutTypes";
 import { RenderModes } from "constants/WidgetConstants";
 import { CanvasWidgetsReduxState } from "reducers/entityReducers/canvasWidgetsReducer";
 import {
-  AlignmentInfo,
   extractAlignmentInfo,
   getAlignmentSizeInfo,
   getStartingPosition,
@@ -15,7 +14,6 @@ import {
   getWrappedRows,
   placeWidgetsWithoutWrap,
   placeWrappedWidgets,
-  Row,
   updateWidgetPositions,
 } from "./positionUtils";
 import { AppPositioningTypes } from "reducers/entityReducers/pageListReducer";
@@ -406,16 +404,7 @@ describe("test PositionUtils methods", () => {
       const result: {
         height: number;
         widgets: CanvasWidgetsReduxState;
-      } = placeWidgetsWithoutWrap(
-        widgets,
-        arr,
-        0,
-        64,
-        false,
-        mainCanvasWidth,
-        columnSpace,
-        0,
-      );
+      } = placeWidgetsWithoutWrap(widgets, arr, 0, false, 0);
       expect(result.height).toEqual(7);
       expect(result.widgets["2"].leftColumn).toEqual(16);
       expect(result.widgets["2"].rightColumn).toEqual(40);
@@ -493,16 +482,7 @@ describe("test PositionUtils methods", () => {
       const result: {
         height: number;
         widgets: CanvasWidgetsReduxState;
-      } = placeWidgetsWithoutWrap(
-        widgets,
-        arr,
-        0,
-        64,
-        false,
-        mainCanvasWidth,
-        columnSpace,
-        0,
-      );
+      } = placeWidgetsWithoutWrap(widgets, arr, 0, false, 0);
       expect(result.height).toEqual(7);
       expect(result.widgets["1"].leftColumn).toEqual(8);
       expect(result.widgets["1"].rightColumn).toEqual(24);
@@ -597,16 +577,7 @@ describe("test PositionUtils methods", () => {
       const result: {
         height: number;
         widgets: CanvasWidgetsReduxState;
-      } = placeWidgetsWithoutWrap(
-        widgets,
-        arr,
-        0,
-        64,
-        true,
-        mainCanvasWidth,
-        columnSpace,
-        0,
-      );
+      } = placeWidgetsWithoutWrap(widgets, arr, 0, true, 0);
       expect(result.height).toEqual(7);
       expect(result.widgets["1"].mobileLeftColumn).toEqual(8);
       expect(result.widgets["1"].mobileRightColumn).toEqual(24);
@@ -696,9 +667,7 @@ describe("test PositionUtils methods", () => {
           ],
         },
         0,
-        64,
         true,
-        mainCanvasWidth,
       );
       expect(result.height).toEqual(18);
       expect(result.widgets["1"].mobileLeftColumn).toEqual(48);
